backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-SocketIO==5.3.6
python-socketio==5.10.0
python-engineio==4.8.0
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
anthropic>=0.39.0
Werkzeug==2.3.7
eventlet==0.33.3
